requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests-oauthlib==1.3.1
SQLAlchemy==2.0.23
python-jose==3.3.0
flask-sqlalchemy==3.1.1
tweepy==4.14.0
python-dotenv==1.0.0
gunicorn==21.2.0
